+++ docs/make.jl
@@ -20,11 +20,13 @@ makedocs(;
         "Home" => "index.md",
         "Modules" => "modules.md",
         "Networks" => "networks.md",
+        "Geographic layers" => "geo.md",
         "Detailed connectivity" => "connectivity.md",
         "Distribution" => "distribution.md",
         "Collections and instances" => "collections.md",
         "Matrices" => "matrices.md",
         "Updates" => "updates.md",
+        "Contingency analysis" => "contingency.md",
         "Interop" => "interop.md",
         "Developer Guides" => [
             "Migrating from 0.10 to 0.11" => "migration-0.11.md",